Move AclCollection to seperate file
